Extract ActionCable object to its own module
This is step in the direction of removing the circular dependencies.
We still have 2, but they have now changed from:

```
(!) Circular dependency: app/javascript/action_cable/index.js -> app/javascript/action_cable/connection.js -> app/javascript/action_cable/index.js
(!) Circular dependency: app/javascript/action_cable/index.js -> app/javascript/action_cable/connection.js -> app/javascript/action_cable/connection_monitor.js -> app/javascript/action_cable/index.js
```

to:

```
(!) Circular dependency: app/javascript/action_cable/action_cable.js -> app/javascript/action_cable/consumer.js -> app/javascript/action_cable/connection.js -> app/javascript/action_cable/action_cable.js
(!) Circular dependency: app/javascript/action_cable/action_cable.js -> app/javascript/action_cable/consumer.js -> app/javascript/action_cable/connection.js -> app/javascript/action_cable/connection_monitor.js -> app/javascript/action_cable/action_cable.js
```

While that might seem like all we did was introduce additional items
in the circular chain, we can now add additional named exports to this
newly added module, and that will allow us to get rid of the circular
dependencies.
